- Removed hardcode, - by price of initializing singleton config class with static fields
